+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added option create duty hours for other people.
 - Added option define `action` from slack message.
 - Added sync missed users from Opsgenie.
+- Added command hide reason. #113
 
 ### Fixes
 - Fixed publishing docker image with latest tag.